Added support to grade multiple datasets at the same time.
Copy the student submissions into the bucket under a folder for that
particular team AND dataset. This way having the separate instances
running does not affect the contents of each other.

Note that this must still be done in a separate shell, where the infile
was set to an alternate value.